+++ traits/GUIDTrait.php
@@ -12,7 +12,7 @@
 
 namespace vistart\Models\traits;
 
-use vistart\Helpers\Number;
+use vistart\helpers\Number;
 
 /**
  * Entity features concerning GUID.